Revert FeatureInfo::Workarounds setup in command buffer.
This is a temporary solution so we can move forward with green gl_tests, until we migrate the setup code from src/content to src/gpu.

Note that in regular chrome codepath, kGpuDriverBugWorkarounds is always passed down to gpu process, so the code path in this CL is never triggered: it is for testing only.

BUG=228979
TEST=gl_tests,gpu_unittests
Review URL: https://codereview.chromium.org/14186003