Fix issue with login items getting recreated.
Added code to track whether chrome has created a login item so we know if it is
safe to delete it and can avoid re-creating it. This flag is also used to do a
better job of determining if a user has deleted the login item.

BUG=140017

Review URL: https://chromiumcodereview.appspot.com/13982009