Remove migration code for rules->state store
(This is about extensioni system.)
For performance reasons, state store is not loaded on browser start-up, but on the first page load.
But the declarative rules need to be activated ASAP, so rules are stored separately from other state information, and the rules store is loaded immediately on start-up.
There was a migration code trying to read rules from the state store as well, but it's ben sitting there for a while, and causes perfomance issues.
Thus we are now removing the migration code.

BUG=166474,176926

Review URL: https://chromiumcodereview.appspot.com/13985010