Refactoring: Extracting UninstallMetrics from GoogleChromeDistributions to separate files.
This is a "low-hanging fruit" to simplify BrowserDistribution. 

The "routines of interest" are:

BuildUninstallMetricsString()
ExtractUninstallMetrics()
ExtractUninstallMetricsFromFile()

These used to be included in GoogleChromeDistribution, but were only called internally (aside from unit tests).

Previously the unit test was grouped with GoogleChromeDistribution, which was jarring owing to use of #ifdef and friend.

There is no change to the internals of the routines of interest, although their locations and visibilities were changed.

BUG=170398
TBR=jhawkins

Review URL: https://chromiumcodereview.appspot.com/14183002