keyboard: Fix a crash when the keyboard is created with a valid text-input type.
The KeyboardController attempts to shwo the keyboard-container after
creation if the current text-client can accept input. However, the
container hasn't been parented yet, and so it ends up causing a crash.
So instead of this, decide the visibility of the keyboard only after
it has been parented.

BUG=none

Review URL: https://codereview.chromium.org/14200002