[Spellcheck] Remove dead IPC message
The DocumentClosed message could never be sent, since the IPC::Sender
is already gone by the time we know we need to send this.

In detail: SpellcheckProvider is a RenderViewObserver. When a
RenderView goes away, it goes through these steps:

1) Call RenderViewObserver::RenderViewGone()
   This sets RenderViewObserver::render_view_ to NULL
2) Call RenderViewObserver::OnDestruct();
3) delete RenderViewObserver()

render_view_ is needed to send an IPC message, but RenderViewGone is
not overrideable. That means we cannot send a notification to the
browser that a RenderView has left the building.

TBR=tsepez@chromium.org
R=rlp@chromium.org
BUG=none

Review URL: https://chromiumcodereview.appspot.com/13966004