cc: Don't do produce/consume on the texture id given in VideoFrame.
This texture id is used for hardware readback to webgl, so producing
the texture id breaks that path.

Instead, add a hardware_resource ResourceId to pass along with the
video frame resources to the VideoLayerImpl. This is a resource id
wrapping the hardware decoder texture. Once we have shared mailboxes
this can be removed, and we can just pass the Mailbox from the
VideoFrame to the video layer.

R=enne,piman
BUG=179729

Review URL: https://chromiumcodereview.appspot.com/14007004