Update GraphingPageCyclerLogProcessor to work with updated Telemetry pagecyclers
The GraphingPageCyclerLogProcessor previously expected that any array in perf
results would be in the form:
[<url1_iter1>, <url2_iter1>, <url1_iter2>, <url2_iter2>, ...]

The updated Telemetry page cyclers will continue to output the overall results
that way, but will also output results by URL. For example:
url1: [<iter1>, <iter2>, ...]
url2: [<iter1>, <iter2>, ...]

This patch causes the processor to process the by-url results in the same way
as other perf results instead of in the page cycler specific manner.

This fixes a bug which caused the process to fail with the following error when
the page cycler update was submitted the first time:
KeyError: 'blog.chromium.org'

The new unittest data in this patch was taken from an actual run of the updated
page cyclers and includes the by-url traces. Without this patch, it fails in the
same way mentioned above.

BUG=None
TEST=./process_log_test.py && ./runtest_annotator_test.py

Review URL: https://codereview.chromium.org/14024008