Split the job queue and the job information in DriveScheduler.
In the current design, the JobInfo structure of running jobs are owned by
the callback of in-flight operations. This makes impossible to access
(= read from UI code or update by progress notification from lower layer)
the information during the job is running.

This patch decouples JobInfo and QueueEntry. QueueEntry is still associated
with the callback function but JobInfo is now owned by the Scheduler itself.

Along the way, JobInfo::job_id became to be filled with the unique job id.

BUG=154243, 128079

Review URL: https://codereview.chromium.org/14251002