Reduce WebRTC buffer size to 40KB
BUG=231705

This is a M27 only change.

This bug is to fix a problem with large buffer size in WebRTC, currently
256KB in M27. This resulted in high estimated RTT and behaves badly with
packet drops and network congestion.

Refactoring is under way in Chrome Canary to reduce this buffer size and
also change the behavior in response to network work congestion. This
choose of 40KB buffer size is tested in Chrome Canary that essentially
has the same buffer size (32 packets). This change is meant to take part
of fix in Chrome Canary and applying it to Chrome M27.

Review URL: https://codereview.chromium.org/14281002