Roll Trace Viewer to r255.
This pulls in the trace viewer changes to switch from include filters
to exclude filters. This will allow unknown categories to be returned
to the trace as we now specify categories we don't want to see instead
of categories we do want to see.

BUG=231608

Review URL: https://codereview.chromium.org/14174004