[NaCl SDK] Pin pepper interfaces in nacl_io.
Because we specify the desired methods directly, there is no drawback to specifying the interface versions.

One huge benefit is being able to build with a newer SDK on an older chrome -- in this case, pinning PPB_FILEREF_INTERFACE to version 1.0 means that we can use nacl_io from the pepper_28 SDK to run on Chrome stable (version 26).

BUG=none
R=noelallen@chromium.org

Review URL: https://codereview.chromium.org/14230011