[Telemetry] Fix race in tab.Navigate's script_to_evaluate_on_commit.
I believe this is what was causing the page cyclers to hang on the bots. We
must add the script to evaluate on load prior to performing the navigation. If
we add it after, it is racy.

The reason it wasn't like this before is that Page.addScriptToEvaluateOnLoad
didn't work on cross-renderer navigations. But that problem is being fixed by
https://codereview.chromium.org/13949006/

The current file-based page cyclers don't do cross renderer navigations, so
will be fine even without the above fix. However, the above fix will be
necessary for the new WPR-based page sets.

BUG=None
TEST=page set based page cyclers on linux
NOTRY=True

Review URL: https://codereview.chromium.org/13926007