History: Workaround bug with clear button in search field.
Due to a Blink bug (http://crbug.com/231830), the clear button in the
search field is not properly rendered if the <header> element has
its transform style property set to anything other than 'none'.

A simple workaround to make it render properly in 99% of cases is to
set the value of -webkit-transform to 'none' instead of 0px.

BUG=231291
TEST=Open chrome://history, type something in the search field, and
ensure that the clear button (a circle with a "X") is visible.

Review URL: https://codereview.chromium.org/14053006