Introduce RawVarData and associated classes for serializing PP_Vars
This provides a replacement for the inner class of SerializedVar for serializing a PP_Var. It achieves 2 purposes: 1) it refactors the design of SerializedVar::Inner which was growing increasingly messy (and was going to become nastier) 2) it provides support for serializing/deserializing dictionary and array PP_Vars. To serialize a PP_Var, first the transitive closure of that var is computed (all nodes referenced) and then each var in the transitive closure is written to the message. Some trickiness arises from the fact that a PP_Var can be either a primitive or a reference and references have to be maintained when transmitting over the wire. The comments in the header files provide a description of the approach in more detail.

BUG=

Review URL: https://codereview.chromium.org/13887007