Fix NOTREACHED triggered by MenuItemView::GetAcceleratorText
A recent change to accessibility on Chrome OS is triggering a call to
MenuItemView::GetAccessibleState on a MenuItemView that has a delegate
but no command id, which asserts in MenuModelAdapter when we try to
look up a command that's not there.

The fix is to not call MenuModelAdapter::GetAccelerator to look up a
command when the command is zero.

BUG=230107,232079

Review URL: https://codereview.chromium.org/14161006