Let Blink try jobs use a Blink-specific LKGR.
Currently we're using mostly the same code for blink try jobs and chromium try jobs. We have hardcoded logic to use the blink HEAD revision for jobs run on the *_layout_* bots, but use the chromium LKGR revision for all bots.

This is not a good configuration to be using. Ultimately we should have a blink LKGR that specifies revisions of both blink and chromium, and have a way to tell try jobs that they are blink changes.

This change is a step along that path, in that we'll start using blink-status.appspot.com/lkgr to pull a chromium revision, which is set manually for now.

Later changes will add a blink-specific lkgr_finder that does something more correct.


R=iannucci@chromium.org
BUG=233418

Review URL: https://chromiumcodereview.appspot.com/14336002