Scale TouchEvent's radius when touchscreen is in mirror mode
TouchEvent's radius on Pixel is reported in the touchscreen's
native resolution. When Pixel is connected to an external screen
and running in mirror mode, the radius value is too large for
the external screen's low display resolution. The radius value
needs to be scaled properly in this case.

BUG=229932
TEST=Switching to mirror mode when connecting Pixel to a low
resolution external display, following the test in the issue
and make sure the radius has a sensible value.

Review URL: https://chromiumcodereview.appspot.com/14279002