Add line height setting to views::Label & use it for notifications.
This fixes the line spacing of notification so it's 18px on all
platforms (instead of the previous 17px on Chrome OS and 15px on
Windows).

To do this, support for a line height property was added to the
BoundedLabel that NotificationView uses, to the Label that BoundedLabel
uses, and to the Canvas that Label uses.

BUG=225871
R=mukai@chromium.org,msw@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14322007