[android] Resize the android_webview if it's 0x0 initially.
This changes the content size update path for android_webview to use
the preferred size RenderView mechanism instead of the
CompositorFrameMetadata.
The reason for the change is due to the fact that the CompositorFrameMetadata
is not updated when the view size is 0x0, which is a common use case
for the WebView when it's layout mode is set to "wrap content".

BUG=b/8187850
TEST=AndroidWebViewTests

Review URL: https://chromiumcodereview.appspot.com/12567020