Re-enable a few find in page tests.
They had been disabled -- not due to inherent flakiness in the test itself, but due to an environment issue (bug 96594) which affects a good number of tests. I'm following the precedent of others by re-enabling these tests while we work on a solution for the underlying environment issue. Note: We've already had a regression due to the tests being disabled.

TBR=sky
BUG=145476, 128724, 152100, 177487

Review URL: https://chromiumcodereview.appspot.com/14295003