Update message center systray icon.
New design shows a notification bubble on a gray background when there
are no unread notifications.  When unread notifications exist, an image
displaying the number of notifications (up to 9+) is rendered on a blue
background.

The new png is a sprite map of the digits.

TEST=manually generate 11 notifications
BUG=189129

Review URL: https://codereview.chromium.org/14192025