Revert r195040 "Let Blink try jobs use a Blink-specific LKGR."
It looks like the tryserver caches the lkgr value into self._last_lkgr
and only updates the value when the new lkgr is newer than the old one.
Since Blink tends to track tip-of-tree closer than Chromium, this means that
we'd use the Blink LKGR most of the time, which would be bad.

TBR=cmp@chromium.org, iannucci@chromium.org
BUG=233418

Review URL: https://codereview.chromium.org/13878026