Initial restyle of the gallery.
This patch introduces very basic restyle, which makes the gallery usable with the new ui. It addes the header which is draggable. This is quite tricky, since the header has to be outside of the iframe. -webkit-app-region: drag has no effect within an iframe.

Along the way, a opacity animation has been added to thumbnails.

TEST=Run with --file-manager-new-ui, click on a photo to get to the gallery. Try to move.
BUG=233485

Review URL: https://codereview.chromium.org/14320020