Instant Extended: Prevent overlay from becoming transparent
If the active tab contents became composited then the overlay would
either become transparent or flicker.

The problem was that both the overlay and active tab contents were
calling SetAllowOverlappingViews(true). This meant that which ever view
drew last would be visible.

Fix was to not call SetAllowOverlappingViews(true) for the overlay. This
means that the find bar will not be visible for the overlay but that's
something we can live with.

This CL also add a new compositingIOSurfaceCreated notificaiton. This
works around a bug where the active tab contents would draw over the
overlay if it was created last. Fix was to re-add the overlay to the
view hiearchy when this happened.

BUG=222122
TEST=
  Make sure "GPU compositing on all pages" is set to "Default".
  Make sure flash is working.
  Navigate to cnn.com.
  Type "http://" in the omnibox. Verify that the overlay is visible.

Review URL: https://chromiumcodereview.appspot.com/14367008