Audio extension API.
Extension API to query and control audio device state.
The design doc is available at

https://docs.google.com/a/google.com/document/d/1szQEkGggTScHP8Ahum8ZrN7QxmHaCsF83XTq2AcsU-g/

This CL is not yet complete, but is compiling and working for
the portion that is finished. It contains most of the stubs
and implements the audio.onDeviceChanged event as well as
part of the audio.getInfo. I would like to land this to trunk
ASAP and hand over the work to Rahul.

BUG=175798
TEST=Trybots, local test app exercising audio API.

Review URL: https://chromiumcodereview.appspot.com/13486004