Never bind two NSOpenGLContexts to an NSView
Binding two NSOpenGLContexts to an NSView will cause corruption and
crashes when one of the contexts is destroyed. Call clearDrawable before
changing the underlying context of a CompositingIOSurface.

Note that the NSView owns the CompositingIOSurface, so there is no risk
of "some other" NSOpenGLContext being bound to the NSView.

BUG=230883

Review URL: https://chromiumcodereview.appspot.com/14092005