Allow test files to use non-platform suffixes for PRESUBMIT checks.
The current check for deciding if a file is a unittest allows
a strict whitelist of platforms as suffixes. However, when a
unittest file becomes too large, it is useful to split it on
other axes than by platform. This allows a _suffix for a topic
to be used without failing production-only presubmit checks.

In other words, it considers foo_bar_unittest_baz.cc to be a
unittest.

There are many examples of foo_unittest_utils.cc which are used
only by unit tests throughout chromium, but they have been lucky
to avoid using FunctionForTesting type functions. In cc/ we have
run into this problem, so allowing our unittests to be considered
non-production code prevents false positive warnings about our
test files.

NOTRY=true
R=maruel@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/14317011