Add ash-enable-sticky-edges for 'sticky' instead of 'snap' behavior.
I renamed everything related to edges to 'sticky' to differentiate
it from SnapSizer in the code. I put the new behavior (which is
a pretty trivial change) behind a flag.

I moved ShouldStickToEdge to window_util.cc in case we want to use
the smae logic in snapping panels to / from the launcher. However,
my initial experimentation with that was not great so I left that
change out for now.

BUG=228955

Review URL: https://codereview.chromium.org/14273008