Explicitly construct/destruct UserManager and dependencies in tests
The ScopedTestUserManager initializes and subsequently tears down the
UserManager singleton only if it has not been initialized already. This
allows ScopedTestUserManagers to be nested but also means that it is
never clear which ScopedTestUserManager actually did end up initializing
the singleton. The same goes for ScopedTestCrosSettings.

This CL makes the initialization and teardown explicit. The helper classes
now explicitly manage the initialization and teardown of their respective
singletons and cannot be nested anymore.

BUG=None

For tests that need to instantiate a UserManager:
TBR=sky,rkc,aa,thestig,asvitkine,davemoore,kinuko,pkotwicz,davemoore,akalin

Review URL: https://codereview.chromium.org/14192040