With this CL we can catch OnResume, OnPause, OnDestroy and OnStop notifications in SimpleCache, so we save our index before we get killed.
OnPause means the app lost focus (went to the background).

The idea is that we can write the index file to disk on every OnPause, and while the app is in the background we can aggressively increase the frequency we write the index file to disk.

BUG=233536

Review URL: https://chromiumcodereview.appspot.com/14362009