Omnibox: Fix HistoryQuick Crash on non-English Locales on Linux
Looking at the function ScoreComponentForMatches(), the only possible
line I see that could have a division by 0 is
  size_t max_significant_length =
      std::min(max_length, std::max(term_length_total, kMaxSignificantLength));
  int complete_value =
      term_length_total * kCompleteMaxValue / max_significant_length;

This change short-circuits the function so that this isn't run when
max_length == 0.

Please say something if you see another line in the function that could
cause a division by 0.

Not tested because I don't know how to reproduce the bug.

TBR=sky
BUG=234096

Review URL: https://chromiumcodereview.appspot.com/14358011