Turn SimpleCache experiment on in Android dev channel.
Specifically, going to 45% in, 45% control group on android dev channel. The 10%
experiment no group will help see bad behaviour from the control group.

We'd like to start running this experiment in this channel soon, and with this CL
we are putting users into the two experiment groups on our target platform. I
now make the experiment the default for all platforms, but leave it 100% off in
those cases.

R=stevet
BUG=229919

Review URL: https://chromiumcodereview.appspot.com/14402010