PNaCl PPAPI shims: Remove function pointer casts for PPP_* shims
The casts are only necessary for the PPB_* shims, where we need to
cast away the "pnaclcall" function attribute.

The PPP_* shim functions must match the default calling conventions
and so do not have this attribute, so removing the casts here provides
a little extra type checking.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3400
TEST=cd ppapi/generators && ./generator.py + compile shims

Review URL: https://codereview.chromium.org/13888011