Touch events converted to cancel should also be put into pending queue
If an event is converted to cancel, it should also be in the pending queue.
Otherwise, when the ack of that touch cancel comes, it will confirm the next touch event.
Adding a boolean to the pending queue to indicate whether an event has been converted to cancel

BUG=234335

Review URL: https://chromiumcodereview.appspot.com/14225018