SyncFS: Determine a file type of a deleted entry by local metadata
So far we treat a file type of a deleted entry (and a trashed entry)
as SYNC_FILE_TYPE_UNKNOWN since the deleted entry does not have a file type.

This patch determines it by local metadata. When a resource ID owned by the
deleted entry is equal to one owned by local metadata, we can assume that a
type of the deleted entry is the same as a type contained by local metadata.

This is a preparation patch for directory operation support.
After this, we can handle a deleted entry along its file type.


BUG=none
TEST=unit_tests

Review URL: https://chromiumcodereview.appspot.com/14211010