Rework how bots get the version of Blink.
The current buildbot code uses a number of hacks to try and get the
correct version of Blink, and has a number of bugs. This patch completely
reworks things.

We will now set a 'blink_config' factory property to determine which
version of blink to use, and pass that to the GClient object. If 'blink_config'
is set to 'chromium' (or None, or anything other than 'blink'), we'll use the
stock gclient spec and the version of Blink as set in the DEPS file.

If 'blink_config' is set to 'blink', then we will use a custom version
of Blink. If the patch contains a "trunk" branch, then we use the revision
in the patch (as the change came from the Blink poller). Otherwise,
we will set the custom "webkit_revision" gclient variable to "", which
is equivalent to HEAD (meaning we'll get the tip-of-tree for Blink).
This replaces the way we used "WK_REV" previously.

The code previously also supported a means to specify a fixed version of
Blink via the revision specified in the patch as uploaded through the
svn poller to Rietveld (in trychange and git-wktry). This code path was
also a hack and we're currently discouraging git-wktry altogether, so
I'm removing this support for now (although I think trychange should still
work properly).

A lot of trouble was caused by us using "inheritance" in different
ChromiumWebKit* helper functions in ChromiumFactory. We should've been using
a parameter instead. This change removes all of the *ChromiumWebKit*
methods and updates all of the bot configs accordingly to use the
"blink_config" parameter as described above.

The net result of all of this is that a bot will now be either a "Blink
tip of tree" bot, or not, and that is independent of the patch being
tried or applied.

R=szager@chromium.org, iannucci@chromium.org
BUG=233412

Review URL: https://chromiumcodereview.appspot.com/14440005