Sync: Turn on full history sync by default.
Full history sync (and history delete directives) were off by default,
but turned on in M27 via an experiment. This meant that history from
signed-in devices was not visible on chrome://history until the browser
was restarted.

This CL make full history sync enabled by default, but retains a
command-line switch and about:flags option to disable it. It also
removes the option to disable delete directives only, since that
had the same effect as disabling full history sync.

TBR=brettw@chromium.org
BUG=233098
TEST=Start up Chrome in a fresh profile, and sign in. Go to
chrome://history and ensure that it says "Showing history from your
signed-in devices" at the top of the history page.

Review URL: https://chromiumcodereview.appspot.com/14344002