Pepper IDL: Check for structs in callbacks.
It's problematic for PNaCl when callbacks take structs that are passed by
value; there's not enough type information for matching GCC's calling
conventions on the target architecture.

This change scans for typedefs that define functions, and scans the arguments
to see if any are a struct (or a typedef of a struct).

This change whitelists one existing problematic callback.

Tested:
  Verified that exceptions are raised for the one type in the whitelist when
the whitelist was empty. Verified that no exceptions are raised with this
change as-is.

BUG=233439

Review URL: https://chromiumcodereview.appspot.com/13973011