Bluetooth: send DeviceChanged for more properties
The address and trusted properties of the underlying D-Bus device
interface correspond to BluetoothDevice properties, so DeviceChanged()
must be sent for a change in value of those too.

In addition the reconnect_mode property of the D-Bus input interface
changing (including the interface being added and removed for this
object) corresponds to BluetoothDevice::IsConnectable() so
DeviceChanged() must also be sent.

BUG=233370
TEST=device_unittests
R=youngki@chromium.org

Review URL: https://codereview.chromium.org/14065018