Make environment for PyAuto tests more customizable.
I know PyAuto is deprecated but we need to keep our tests a live just a
little bit longer.

This is needed to resolve the differences between Mac and Linux. My
previous fix in r196075 didn't work out well as I only tested on Linux.
It started causing errors on Mac execution, so I thought it was
better to move the configuration out into the config files instead.

As I was the one introducing the env parameter to
AddPyAutoFunctionalTest in r174380, I thought it was cleaner to remove
it again and make the environment configuration using a factory
properties key instead.

TEST=local build on Mac and Linux
BUG=none

Review URL: https://chromiumcodereview.appspot.com/14447005