Improve FrameSubscriber decision-making for smoother frame capture sampling.
Added a "present time" argument to the ShouldCaptureFrame() method to communicate to the implementation when a frame is to be displayed on the screen (e.g., around regular v-sync intervals).  Then, replaced the existing sampling logic with use of a token bucket.

This is needed to support smooth frame capturing on Mac when the asynchronous read-back code path is used (in CompositingIOSurfaceMac).  See: https://codereview.chromium.org/14361007/  This change is also likely to produce none to some improvement on other platforms.

BUG=232006,232264,232256

Review URL: https://chromiumcodereview.appspot.com/14384002