Prevent chrome.app JSON schema from loading on every page
The app API along with app.window and app.runtime have been converted to use
the feature system. Bindings are not added to the chrome object for unavailable
APIs that are children of available APIs. For example, if chrome.app is
available, we will not add lazy bindings to chrome for app.window and
app.runtime. This eliminates the need to load the app schema, because we no
longer need to get chrome.app to add the app.runtime and app.window bindings.

BUG=55316

Review URL: https://chromiumcodereview.appspot.com/13604005