Views Textfield: Don't start drags during multiple clicks
This change is to match the behavior on Win where you can perform a multiple click, do not release the button, drag, and it performs a selection.

The view.cc change is to make sure OnMouseDragged is called when there is a possible drag but it cannot be started (i.e GetDragOperationsForView returns something but CanStartDragForView doesn't allow).

BUG=234362

Review URL: https://chromiumcodereview.appspot.com/14236009