[Android WebView] Move the most of WebSettings into AwSettings
Historically, WebSettings management is used to be in the ContentSettings class,
while only WebView actually needs it. Chrome for Android is only interested in
the value of the "JavaScript enabled" setting.

This change leaves zoom-related settings intact, as moving them will require
decoupling of the ZoomManager class from the ContentView* family of classes,
which deserves a separate change.

BUG=b/8296421

Review URL: https://chromiumcodereview.appspot.com/14271024