Utility word trimmer for sending feedback to spellcheck server
This CL adds a utility word trimmer to be used in sending feedback to the
spellcheck service. The feedback will include several words of context around
the misspelled word. This utility will be responsible for trimming the
spellchecked text to include only a certain number of words of context.

TBR=jhawkins
BUG=170514

Review URL: https://chromiumcodereview.appspot.com/14489003