Add transition animations to notification popups.
This adds following transitions to the popups (aka toasts):
- new notification slides from the right into position
- expired or closed notification slides to the right
- after the closed notificaiton slides out, a stack slides down 
  appropriately, after that if there is a waiting notification it slides in
- if user starts to close notifications by clicking X button, the stack 
  goes into "user clicking X" mode in which the remaining popups are sliding
  under the mouse pointer, but new notifications do not enter. The mode ends
  in about 2 seconds after user stopped closing, with remaining of the stack
  sliding to its regular place and then new notifications entering.

I refactored a little the MessagePopoupCollection class, now it does not keep
the list of both Widgets and Toasts, but only a list of Toasts which are
delegates of their Widgets.

The logic includes sequencing of operations (close->sliding->adding new ones)
because when all things are animated at the same time it creates
even worse effect then just instant update.

The test can not be done the same way, because it now has to wait for
sequences of operations. I've updated a test to do what I think is a good
thing to do (to test the actual code path). I plan to do a followup CL
to reduce the time of aniimations in test to near 0 and add more tests.

BUG=233660

Review URL: https://chromiumcodereview.appspot.com/13996021