Bluetooth: Display a "connecting" label next to the device.
The list of Bluetooth devices in the settings page does not indicate
if the Bluetooth device is connecting or not. In the case of a
connecting device, it will appear as connected or not connected, without
an indication that a connection action is beign performed in the device.

This patch shows a "(Connecting)" message next to it. Also includes some
minor documentation fixes.

BUG=220554
TEST=Manual test

Manual test procedure:
1. Open a chrome://settings page and expand the "advanced settings" part.
2. While keeping the list of bluetooth devices there on the screen, connect
to a new bluetooth device from the system tray. While pairing it using the
system tray, the device should appear as "(Connecting)" in the settings
page.

Review URL: https://chromiumcodereview.appspot.com/14146012