aura: make cc root layer clip descendants
The cc root layer (just above the ui root layer) is set to the size of the
screen, and setting clipping is a visual noop. However it makes sure render
surfaces are clipped to the size of the screen, when they could be created
bigger because of scale.

BUG=177256

Review URL: https://chromiumcodereview.appspot.com/13982016