Add ability to sign-out of Google directly from profile switcher menu.
As part of the plan to make sync sign-in more integral to a user's profile, we're adding an explicit ability to "sign out" of the profile in place of signing out of Google inside the content area.  This CL adds that basic sign-out functionality by adding a "sign out" button to the avatar menu that, when clicked, goes through the steps of signing out of Google and closing all browser windows associated with that profile.

Future CLs will do the Right Thing(tm) of requiring sign-in before opening that profile so that the tabs can resume where there were, just as though there had been no sign-out event.

The UI included here is temporary.  A new profile-switcher menu is being discussed and the core functionality included here will eventually get triggered from that once the design is finalized.  In other words, the UI presented here will eventually be replaced.

See it in action:
https://docs.google.com/a/google.com/file/d/0B4S7cIAPVn9XdzJFZXl2SncxMlE/edit?usp=sharing

Previous avatar menu:
https://docs.google.com/a/google.com/file/d/0B4S7cIAPVn9XY3dZMXRUNGtHd3M/edit?usp=sharing

To test:
- launch with --enable-profile-signout
- create a second profile (so Avatar menu appears)
- sign-in to a Google account (like gmail)
- open Avatar menu and click the "sign out" button
- (window will close when sign-out is complete)
- re-open profile
- go to a Google account
- note that you are not signed in

BUG=

Review URL: https://chromiumcodereview.appspot.com/12726012