initial pass at the steps to use for blink in lkgr_finder.
This patch adds the initial list of bots and steps to use to
determine an LKGR for Blink. Blink's LKGR is defined (roughly) as
the most recent version of Chromium that compiles for all of the configurations
used on the chromium.webkit waterfall. Note that this is a *Chromium*
revision, not a *Blink* revision.

R=xusydoc@chromium.org
BUG=233418

Review URL: https://chromiumcodereview.appspot.com/14031022