Webapp changes to support third party authentication
This uses an OAuth flow on the server to fetch the token and shared secret. There are two implementations for this:

* The current one manually opens a tab and asks for a redirect to a blank page in talkgadget, which we content-script to sendmessage the token/secret back to the extension (fairly similar to our OAuth trampoline)

* Once we're running on appsv2, and identity is out of experimental, we can use launchWebAuthFlow to do this.

This includes an interstitial to ask for an optional permission to the given host. The window.open method doesn't actually need this, but the identity API one does, so I thought I'd leave it in to make its behavior match closely the one of the identity API, which is the one we'll use in the future.

Most of the code is shared between these two versions, the only different pieces are the mechanics to open the window/launchWebFlow, and to send the redirectedUrl back to the webapp for parsing.

BUG=115899

Review URL: https://chromiumcodereview.appspot.com/12905012