Enable receiving a list of markers in text with spellcheck requests
This CL enables Blink to send a list of markers together with text to be
spellchecked. The CL is structured such that Blink can either send the markers
or not, depending on Blink version. Once Blink dependency is rolled, I will
address the TODOs in the CL.

BUG=170514

Review URL: https://chromiumcodereview.appspot.com/13863016