Make sure pnacl_error_handling_test's gyp "output" is created.
Otherwise, it will always try to build the target, even
though it has already been built.  Since we used
build_nexe.py's --compile, we have to set the output name to
be the object file instead of the pexe.

Alternatively, we could use a separate python script
to build this magical incomplete "pexe".  Or, we can
look for a different way to test linker failures.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3364

Review URL: https://chromiumcodereview.appspot.com/14391006