Fix crash in ToastContentsView::StartTimer()
The crash is caused by notifications that are already expired when the user moves the mouse off them. This causes the toast to be removed immediately and synchronously from the toasts_ collection which kills the iterator.
Instead of modifying the iterator loop, I'm making the actual call to CloaseWithAnimation() always async rather then have it sync in some cases and async in others.

BUG=235908

Review URL: https://codereview.chromium.org/14283008